Move drop code from actions/drag to actions/drop
